+++ htdocs/modules/debugbar/admin/permissions.php
@@ -18,7 +18,7 @@
 
 use Xoops\Core\Request;
 use Xmf\Module\Helper;
-use Xmf\Module\Permission;
+use Xmf\Module\Helper\Permission;
 
 include_once __DIR__ . '/admin_header.php';
