+++ htdocs/modules/monolog/admin/permissions.php
@@ -19,7 +19,7 @@
 
 use Xoops\Core\Request;
 use Xmf\Module\Helper;
-use Xmf\Module\Permission;
+use Xmf\Module\Helper\Permission;
 
 include_once __DIR__ . '/admin_header.php';
